Refactor enclave builder code
