+++ enclave-manager/web/packages/app/src/emui/enclaves/components/enclaveBuilder/KurtosisServiceNode.tsx
@@ -1,13 +1,13 @@
 import { Flex, Tab, TabList, TabPanel, TabPanels, Tabs } from "@chakra-ui/react";
 import { memo } from "react";
 import { NodeProps } from "reactflow";
-import { BooleanArgumentInput } from "../../form/BooleanArgumentInput";
-import { DictArgumentInput } from "../../form/DictArgumentInput";
-import { IntegerArgumentInput } from "../../form/IntegerArgumentInput";
-import { KurtosisFormControl } from "../../form/KurtosisFormControl";
-import { ListArgumentInput } from "../../form/ListArgumentInput";
-import { StringArgumentInput } from "../../form/StringArgumentInput";
-import { KurtosisFormInputProps } from "../../form/types";
+import { BooleanArgumentInput } from "../form/BooleanArgumentInput";
+import { DictArgumentInput } from "../form/DictArgumentInput";
+import { IntegerArgumentInput } from "../form/IntegerArgumentInput";
+import { KurtosisFormControl } from "../form/KurtosisFormControl";
+import { ListArgumentInput } from "../form/ListArgumentInput";
+import { StringArgumentInput } from "../form/StringArgumentInput";
+import { KurtosisFormInputProps } from "../form/types";
 import { MentionStringArgumentInput } from "./input/MentionStringArgumentInput";
 import { MountArtifactFileInput } from "./input/MountArtifactFileInput";
 import { PortConfigurationField } from "./input/PortConfigurationInput";
