+++ enclave-manager/web/packages/app/src/emui/enclaves/components/enclaveBuilder/VariableContextProvider.tsx
