+++ enclave-manager/web/packages/app/src/emui/enclaves/components/enclaveBuilder/input/FileTreeArgumentInput.tsx
@@ -3,7 +3,7 @@ import { FileTree, FileTreeNode, isDefined } from "kurtosis-ui-components";
 import { useMemo, useState } from "react";
 import { Controller } from "react-hook-form";
 import { FiPlus } from "react-icons/fi";
-import { KurtosisFormInputProps } from "../../../form/types";
+import { KurtosisFormInputProps } from "../../form/types";
 import { EditFileModal } from "../modals/EditFileModal";
 import { NewFileModal } from "../modals/NewFileModal";
